Various polishes on Modal refactoring
- use overlayScrollbar from electron blink features, to prevent
scrollbar taking space. as a side effect we can now scroll by holding
scrollbar with mouse left button. this is a native behavior that we
should not decide to remove for users
- update bottom white gradient to prevent overlapping scrollbar
- make account settings modal scrollable, but increase it's default
height so no weird Select triggering scroll when opening
- remove GrowScroll from release notes modal so we don't have two nested
scrolls (was producing really weird behaviour, scrollbar thumb not
indicating scroll % properly)